+++ 03-next-travel/src/components/boards-list/search/types.ts
@@ -1,0 +1,10 @@
+import { ChangeEvent } from "react";
+
+export interface SearchInputProps {
+    onSearch: (keyword: string) => void;
+}
+
+export interface UseSearchReturn {
+    keyword: string;
+    onChangeKeyword: (event: ChangeEvent<HTMLInputElement>) => void;
+}
